Now when running buildsite, the cogbin will be updated.

+++ buildsite.sh
@@ -1,14 +1,18 @@
 #!/bin/bash
 
-#TODO: Need to make a virtualenv that has the latest sphinx, along with all the needed packages for the various versions of the docs.
 #TODO: Need to make a 'makebook' function
 #TODO: Need to silence sphinx
 #TODO: Add search box on site template. Include doc page inventories for searching
 
-SRC=`pwd`
+REPOROOT=${HOME}/siterepo
+SITEREPOROOT=${REPOROOT}/website
+TGDOCS=${REPOROOT}/tg2docs
+
 WORK=${HOME}/tgsiteworkspace
 SITE=${HOME}/tgsite
-TGDOCS=${HOME}/tgdocs
+VENVROOT=${HOME}/virtualenvs
+
+TGSITEURL=https://github.com/TurboGears/tgwebsite.git
 TGDOCSURL=https://github.com/TurboGears/tg2docs.git
 
 function sitesync() {
@@ -18,9 +22,32 @@ function sitesync() {
     rsync -a ${OPTS} ${SRC} ${DEST}
 }
 
+function mkvenv() {
+    NAME=$1
+    test -e ${VENVROOT}/${NAME} || virtualenv --no-site-packages ${VENVROOT}/${NAME}
+}
+
+function venvon() {
+    VENV=$1
+    source ${VENVROOT}/${VENV}/bin/activate
+}
+
+function venvoff() {
+    deactivate
+}
+
+function invenv() {
+    VENV=shift
+    CMD=shift
+
+    venvon ${VENV}
+    ${CMD} $*
+    venvoff
+}
+
 function checkcommands() {
     PROBLEMS=""
-    for i in git make rsync ; do
+    for i in git make rsync virtualenv sphinx-build; do
 	which ${i} >/dev/null 2>&1
 	if [ $? -ne 0 ]; then
 	    echo Missing command: ${i}
@@ -35,27 +62,38 @@ function checkcommands() {
 
 function workinit() {
     test -e ${WORK} && exit 1
+    test -e ${REPOROOT} || mkdir -p ${REPOROOT}
     test -e ${SITE} || mkdir -p ${SITE}
-    mkdir -p ${WORK}
-    cd ${SRC}
-    git pull
+    test -e ${SITEREPOROOT} || git clone ${TGSITEURL} ${SITEREPOROOT}
     test -e ${TGDOCS} || git clone ${TGDOCSURL} ${TGDOCS}
+    
+    mkdir -p ${WORK}
+    
+    cd ${SITEREPOROOT}
+    git pull
     cd ${TGDOCS}
     git pull
 }
 
 function makesitehtml() {
-    cd ${SRC}/src
-    rm -rf _build
-    mkdir _build
-    touch _build/.hidden
+    mkvenv sitebuild
+    venvon sitebuild
+    pip install --upgrade sphinx
+    cd ${SITEREPOROOT}/src
+    test -e _build && rm -rf _build
+    ${SITEREPOROOT}/bin/cogbin.py
     make html
-    sitesync ${SRC}/src/_build/html/ ${WORK}/
+    git checkout cogbin.rst
+    sitesync ${SITEREPOROOT}/src/_build/html/ ${WORK}/
+    venvoff
 }
 
 function maketg2docbranch() {
     BRANCH=$1
     OUTLOC=${WORK}/$2
+    mkvenv docsbuild
+    venvon docsbuild
+    pip install --upgrade sqlalchemy python_memcached tgext.geo mapfish sphinx
     cd ${TGDOCS}
     git checkout ${BRANCH}
     cd docs
@@ -64,6 +102,7 @@ function maketg2docbranch() {
     test -e ${OUTLOC} || mkdir -p ${OUTLOC}
     sitesync _build/html/ ${OUTLOC}/
     rm -rf _build/html
+    venvoff
 }
 
 function finalize() {
@@ -79,14 +118,16 @@ maketg2docbranch a025e26483fcf5cdc800bd4d8bcac9ee290ef0a7 2.0/docs
 maketg2docbranch tg2.1.5 2.1/docs
 maketg2docbranch development 2.2/docs
 
-# sync 1.0 to top of workspace
-# sync 1.1 to top of workspace
-# sync 1.5 to top of workspace
+# sync 1.0/docs to top of workspace
+# sync 1.0/downloads to top of workspace
+# sync 1.1/docs to top of workspace
+# sync 1.1/downloads to top of workspace
+# sync 1.5/docs to top of workspace
+# sync 1.5/downloads to top of workspace
 # sync 2.0/downloads to top of workspace
 # sync 2.1/downloads to top of workspace
 # sync 2.2/downloads to top of workspace
 # sync remaining static files to top of workspace
-# generate cogbin and sync to top of workspace
 # generate planet and sync to top of workspace
 
 finalize
